Move type_registry and type_registry_state to catalog.{hpp,cpp}
