+++ libvast/test/system/exporter.cpp
@@ -14,10 +14,10 @@
 #include "vast/concept/parseable/vast/expression.hpp"
 #include "vast/detail/spawn_container_source.hpp"
 #include "vast/query_options.hpp"
+#include "vast/system/catalog.hpp"
 #include "vast/system/importer.hpp"
 #include "vast/system/index.hpp"
 #include "vast/system/posix_filesystem.hpp"
-#include "vast/system/type_registry.hpp"
 #include "vast/table_slice.hpp"
 #include "vast/test/fixtures/actor_system_and_events.hpp"
 #include "vast/test/fixtures/table_slices.hpp"
