+++ libvast/test/system/importer.cpp
@@ -16,8 +16,8 @@
 #include "vast/detail/spawn_container_source.hpp"
 #include "vast/format/zeek.hpp"
 #include "vast/system/archive.hpp"
+#include "vast/system/catalog.hpp"
 #include "vast/system/source.hpp"
-#include "vast/system/type_registry.hpp"
 #include "vast/table_slice.hpp"
 #include "vast/test/fixtures/actor_system_and_events.hpp"
 #include "vast/test/fixtures/table_slices.hpp"
